Refactor hub related stuff from session utils
